Minor ts fixes
