+++ src/pages/courses/[id].tsx
@@ -15,8 +15,6 @@ type Props = { course: Data<CourseWithLecturesAndBlocks> }
 export default function CoursePage({ course }: Props) {
   const [showLectures, setShowLectures] = useState(false)
 
-  const handlePptxDownload = () => {}
-
   return (
     <LearningMaterialContainer>
       <LearningMaterialOverview id="source-html">
@@ -50,7 +48,6 @@ export default function CoursePage({ course }: Props) {
           title: course.attributes.Title,
           courseId: course.id,
         }}
-        handlePptxDownload={handlePptxDownload}
       ></MetaDataContainer>
     </LearningMaterialContainer>
   )
